initial commit in new repo
fixed all routes/file paths

fixed "map is not a function error" received once user went back to the search page from an individual movie page. Fix: State is originally defined as an empty array. When it is updated for the single movie, however, it becomes just an object. So when they travel back to search, state is still an object and map doesn't work on objects. To fix, wrapped the return statement in the movieSlice file in [ ] and changed the paths to data in the individual movie page to reflect.

Joe suggested that I make the selected movie a different state object, make a new slice file for it, to avoid that. Makes more organizational sense.